Refactor(CopyrightFooter): Use theme provider for background color
This commit updates the `CopyrightFooter` widget to use the newly introduced `themeProvider` for setting its background color.

Previously, the background color was hard-coded as `Colors.black87`. This has been replaced with `theme.darkBackground` from the `ThemeModel` to ensure that the footer's appearance is consistent with the overall application theme.

This change improves theme adherence and makes the widget more maintainable and scalable.